Extract 'Compound' class into separate module
